Improve Gravity, add GUI
Improve Gravity - dimensions are now correct, so it is responding to the correct underlying pixel.
Add GUI to give feedback to turn on or off the various render features

+++ src/gnome.cpp
@@ -92,7 +92,7 @@ string gnome::chooseRandomGnome() {
 //--------------------------------------------------------------
 void gnome::setRandomPosition() {
     x = 0;
-    y = ofRandom(h/2, 1080-(h/2));
+    y = (int)(ofRandom(h/2, 1080-(h/2)));
     
     // 50% chance to flip horixontally
     if (ofRandom(2) > 1) {
